requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pandas
pypdf
python-docx
spacy
nltk
phonenumbers
tqdm